refactor: isolate document source memoization
